organize the components
