+++ src/components/LabeledValue/LabeledValue.css
