+++ src/components/LabeledValue/index.js
