+++ src/components/TemplatePreview/TemplatePreview.css
